# apps/desktop/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "eigenheim engine",
    "version": "0.1.0"
  },
  "paths": {
...
    "/datasources/posthog/test": {
      "post": {
        "summary": "Posthog Test",
        "operationId": "posthog_test_datasources_posthog_test_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PosthogIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosthogTestOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/datasources/posthog/sync": {
      "post": {
        "summary": "Posthog Sync",
        "operationId": "posthog_sync_datasources_posthog_sync_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PosthogIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosthogSyncOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/decisions": {
      "get": {
        "summary": "List Decisions",
        "description": "List all decisions, newest first.",
        "operationId": "list_decisions_decisions_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DecisionListItemOut"
                  },
                  "title": "Response List Decisions Decisions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Decision",
        "description": "Create a decision, optionally capturing the current values of listed metrics.",
        "operationId": "create_decision_decisions_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/decisions/{decision_id}": {
      "get": {
        "summary": "Get Decision",
        "description": "Get one decision with its full captured metric snapshot.",
        "operationId": "get_decision_decisions__decision_id__get",
        "parameters": [
          {
            "name": "decision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Decision Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Decision",
        "description": "Partial update of a decision's title, rationale, or status.",
        "operationId": "update_decision_decisions__decision_id__patch",
        "parameters": [
          {
            "name": "decision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Decision Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionUpdateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Decision",
        "description": "Delete a decision and all its captured metric snapshot rows.",
        "operationId": "delete_decision_decisions__decision_id__delete",
        "parameters": [
          {
            "name": "decision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Decision Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DecisionIn": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale",
            "default": ""
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": ""
          },
          "metric_logic_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Metric Logic Ids",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "DecisionIn"
      },
      "DecisionListItemOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "rationale",
          "status",
          "created_at"
        ],
        "title": "DecisionListItemOut"
      },
      "DecisionMetricOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "decision_id": {
            "type": "string",
            "title": "Decision Id"
          },
          "logic_id": {
            "type": "string",
            "title": "Logic Id"
          },
          "captured_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Captured Value"
          },
          "trace_ref": {
            "type": "string",
            "title": "Trace Ref"
          },
          "captured_at": {
            "type": "string",
            "title": "Captured At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "decision_id",
          "logic_id",
          "trace_ref",
          "captured_at"
        ],
        "title": "DecisionMetricOut"
      },
      "DecisionOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/DecisionMetricOut"
            },
            "type": "array",
            "title": "Metrics"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "rationale",
          "status",
          "created_at",
          "metrics"
        ],
        "title": "DecisionOut"
      },
      "DecisionUpdateIn": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "title": "DecisionUpdateIn"
      },
      "DeletedOut": {
        "properties": {
          "deleted": {
            "type": "string",
            "title": "Deleted"
          }
        },
        "type": "object",
        "required": [
          "deleted"
        ],
        "title": "DeletedOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PosthogIn": {
        "properties": {
          "host": {
            "type": "string",
            "title": "Host",
            "default": "eu"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key"
          },
          "days": {
            "type": "integer",
            "title": "Days",
            "default": 30
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "api_key"
        ],
        "title": "PosthogIn"
      },
      "PosthogSyncOut": {
        "properties": {
          "ingested": {
            "type": "integer",
            "title": "Ingested"
          }
        },
        "type": "object",
        "required": [
          "ingested"
        ],
        "title": "PosthogSyncOut"
      },
      "PosthogTestOut": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "events_visible": {
            "type": "integer",
            "title": "Events Visible"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "events_visible"
        ],
        "title": "PosthogTestOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}